implement UI response to authorization

+++ backend/routes/posts.js
@@ -48,8 +48,10 @@ router.post("", checkAuth, multer({ storage: storage }).single('image'), (req, r
     title: req.body.title,
     content: req.body.content,
     // construct path to the image
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
+
   // save to the mongodb
   // the collection will be the plural of upur model
   post.save().then(createdPost => {
@@ -81,12 +83,17 @@ router.put("/:id", checkAuth, multer({ storage: storage }).single('image'), (req
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   });
-  Post.updateOne({_id: req.params.id}, post)
+  // automatically update post if id AND creator matches
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post)
     .then(result => {
-      console.log(result);
-      res.status(200).json({message: "Update successful!"})
+      if (result.nModified > 0) {
+        res.status(200).json({message: "Update successful!"});
+      } else {
+        res.status(401).json({ message: "You are not authorized to update the post!"});
+      }
     });
 });
 
@@ -139,9 +146,14 @@ router.get("/:id", (req, res, next) => {
 
 router.delete("/:id", checkAuth, (req, res, next) => {
   // console.log(req.params.id);
-  Post.deleteOne({_id: req.params.id}).then(result => {
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
     console.log(result);
-    res.status(200).json({message: 'Post deleted'});
+    console.log(result.nModified);
+    if (result.n > 0) {
+      res.status(200).json({ message: 'Post deleted' });
+    } else {
+      res.status(401).json({ message: 'You are not authorized to delete the post.' });
+    }
   })
 });
